split single file into separate file
