+++ pystac/__init__.py
@@ -9,6 +9,7 @@ from pystac.link import (Link, LinkType)
 from pystac.catalog import (Catalog, CatalogType)
 from pystac.collection import (Collection, Extent, SpatialExtent, TemporalExtent, Provider)
 from pystac.item import (Item, Asset)
+from pystac.item_collection import ItemCollection
 from pystac.eo import *
 from pystac.label import *
